+++ CHANGELOG.md
@@ -43,7 +43,8 @@
 - **Global**. Added support HTML intellisense in VSCode: follow [instructions](README.md#installing--usage) to use this
 - **Controls**
   - `button[clear]` not hidden anymore for required controls (user must have ability to clear all at once & put new text). To rollback it use css-rule `wup-text[required] button[clear] { display: none; }`
-  - update hover & focus tyles
+  - `button[clear]` visible only on focus & hover. To rollback it use css `wup-text button[clear] { display: block!important; }`
+  - updated hover & focus tyles
 - **Combobox controls (Select, Date, Time)**
   - Now menu is hidden by default when `autoFocus` enabled. To revert to previous behavior use `WUPSelectControl.$defaults.showCase |= ShowCases.onFocusAuto`
 - [SelectControl](src/controls/select.ts).[SelectManyControl](src/controls/selectMany.ts).